LPS-163245 Cascade rule only applies to the parent relationship

+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/related/models/ObjectEntryMtoMObjectRelatedModelsProviderImpl.java
@@ -83,7 +83,8 @@ public class ObjectEntryMtoMObjectRelatedModelsProviderImpl
 
 		if (Objects.equals(
 				objectRelationship.getDeletionType(),
-				ObjectRelationshipConstants.DELETION_TYPE_CASCADE)) {
+				ObjectRelationshipConstants.DELETION_TYPE_CASCADE) &&
+			!objectRelationship.isReverse()) {
 
 			for (ObjectEntry objectEntry : relatedModels) {
 				_objectEntryLocalService.deleteObjectEntry(
